Getting rid of some TODO's
The PriorityQueue had a TODO checking for null of a generic type,. Since
the Remove method is only used by the Transaction.Close method, and it
has a guard clause checking for empty, we can remove the null check in
the PriorityQueue.

The TODO on the FilteredEventSink was removed since the EventSinkRunner
class has been removed (reducing the smellyness), and because the class
has some logic in it that's worth the value of the class existing.