IDisposable added to Behavior,Event,&Listener
Since none of the Stop methods have a corresponding Start method, the
Stop methods are effectively destructors. The obvious choice to make it
clear that they cannot be restarted is to just call Stop Dispose,
bringing us back to the Disposable Pattern.

http://msdn.microsoft.com/en-us/library/fs2xkftw(v=vs.110).aspx

IDisposable is typically used to free unmanaged resources when they are
no longer needed. In this case, we have no unmanaged resources. However,
IDisposable is a nice and simple interface that makes the intent clear.